pyproject.toml
[project]
name = "aie6-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "chainlit>=2.5.5",
    "cohere>=5.15.0",
    "faiss-cpu>=1.11.0",
    "ipykernel>=6.29.5",
    "langchain-cohere>=0.4.4",
    "langchain-huggingface>=0.2.0",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.4.3",
    "llama-index>=0.12.34",
    "llama-index-vector-stores-qdrant>=0.6.0",
    "lxml>=5.4.0",
    "nltk==3.9.1",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "pymupdf>=1.25.5",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-pptx==1.0.2",
    "qdrant-client>=1.14.2",
    "ragas==0.2.10",
    "sqlalchemy>=2.0.40",
    "unstructured>=0.17.2",
    "websocket>=0.2.1",
    "websockets>=15.0.1",
]
